Move DashboardScreen to a separate directory
