+++ src/screens/DashboardScreen/DashboardScreen.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Text, View } from 'react-native';
 
-import { NavigationData } from '../navigation/AppNavigator/DashboardNavigator';
+import { NavigationData } from '../../navigation/AppNavigator/DashboardNavigator';
 
 type Props = NavigationData<'Dashboard'>;
